Reorganize metal and vulkan backend files
